Split logger tests into focused test files and improve quality
Break the 715-line monolithic logger.test.ts into 5 focused files
in test/lib/logger/:
- formatting.test.ts — errorCodeLabel, formatErrorMessage, formatRequestError
- log-error.test.ts — logError, logErrorLocal, activity log persistence
- log-output.test.ts — logRequest, logDebug, createRequestTimer
- redact-path.test.ts — redactPath pure function tests
- request-id.test.ts — runWithRequestId, getRequestId

Quality improvements:
- Remove tautological ErrorCode constants test (asserted values just set)
- Remove 2 redundant logDebug category tests (kept meaningful variant)
- Replace flaky setTimeout-based timer test with deterministic assertion
- Add new test: "skips ntfy and activity log outside pending work scope"
- Add new test: "all labels are non-empty strings" for errorCodeLabel
- Add new test: "ignores eventId" for formatErrorMessage
- Add new test: "redacts path in output" for formatRequestError
- Add expectLogged helper to reduce logRequest test boilerplate
- Fix env var cleanup with proper try/finally
- Wrap logError + logErrorLocal in outer describe for sequential execution
  to prevent AsyncLocalStorage context leakage
